REVISIT LATER - Revert "Should we minimize "reduction" operator?"
This reverts commit e1b6b7130f4d54ebda7b847ea1fece3b23c7acfa.

It is more urgent to produce a good automaton rather than having an
efficient compiler.
I changed my mind for now and decided to not specialize the reduction
operator and to just minimize the final automaton, even if it means a
lot of work is duplicated.

Hopefully, the specialization should still speed up the compiler by
avoiding some combinatorial explosion very early.