WIP Coverage check
The current approach is too complicated.  The base check is for sure the
"Lrc / Offering_stacks" enumerator.

Extension, such as "Error_stacks" enumerator, have to build on top of it
(because we only want to explore valid stacks). However, the way it is
done now is unsatisfying on many levels:
- we mixed error detection and offering stacks logic
- we sometime explore state without knowing whether they are reachable
  or not, and so not only do we explore more than necessary but we also
  have a logic to validate an error after finding it by checking that it
  is indeed reachable

A better approach would be to explicitly represent enumerators as NFAs:
- keep offering stacks as it is, model other enumerators on it
- rewrite error stacks to be a subset offering stacks with the same
  interface