About to revert the recent work on unreduction graph...
The problem is epsilon-transitions (sequence of reductions that do not
shift any token, with leaves of the form A ::= \epsilon).

To deal with them, we have to compute a form of closure on the graph
that is already solved by the Redgraph.

So in the end, it is maybe a better startint point.