Split disjunctions before simulating reduction: maximize reuse
Simulating reductions starts from a set of continuations.
There is a tension between taking the simulation of union of
continuations, or taking the union of the simulation of each
continuation.
E.g. since reduce (A U B) = (reduce A) U (reduce B), how should we
interpret that?

The original code was doing the union before the simulation. It
results in less derivations of bigger terms.  This was chosen because
if only a single reduction is needed, it requires a bit less work.

This patch changes that: the input is split into individual
continuations that are reduced independently. It usually means more
derivation of smaller terms. But if the same term appears in multiple
reductions, more work can be shared.
This turns out to be faster for realistic workloads (at least the
OCaml grammar :-)).

A second benefit is that it is easier to track individual reductions,
which will simplify dealing with non-determinism during register
allocation.