Filtering total clauses dynamically
Doing so dynamically rather than statically leads to a smaller automaton.

Actually, it has less transitions and states before minimization, but
slightly more states after minimization.

Getting the best of both worlds is possible by restricting the parts of
the automaton to force based on the min and max clauses.