Study influence of transition ordering on chain computation
Dynamic chain computation is important for compressing the automaton.
It currently relies on a heuristic that is sensitive to the ordering and
factorization of transitions.

The first surprise was the significant difference in the size of the
compiled automaton: depending on the ordering, the automaton can go from
180k to 300k, with the longest chain varying between 3 to 6 elements.
This deserves further study.

This came when testing two different ways to compute transitions, which
can be tested by comparing:
$ dune clean; dune build
$ dune clean; FINE_MERGING=1 dune build

The good news is that both automata seems semantically equivalent (the
testsuite pass of course, but also the statistics computed by the chain
generation algorithm are the same).