Refactor reduction graphs
Explicitly represent initial states to make some treatments more
uniform.

The number of enumeration is also much smaller... This is actually a bug
that I fixed accidently: initial states weren't considered when
closing the various relations on lookahead reachability
(potential_reject_after, before, etc).

This led to many more paths being explored because the lookaheads
leading to immediate acceptance or rejectance led to a deep exploration
to find other what else could happen... And this is a very common case.