Redgraph: distinguish states reached immediately after iniating a reduction (fix #12)
The redgraph models reductions that are possible based on a known suffix
of the stack.
A node is a stack suffix and an edge is a reduction that relate two
suffixes.

Initial states are slightly different: since we know nothing about the
stack, we have to peek at the top state to determine which reductions
are possible.
After following any reduction that consume this top state, the next
stack suffixes will be made only of known non-terminals, the target of
the goto transitions that were reached.

TODO: It is likely possible to avoid this irregularity by changing the
construction of the redgraph, but I did not find a way yet.